+++ db/migrate/20151124212941_add_labels_to_variables.rb
@@ -1,0 +1,5 @@
+class AddLabelsToVariables < ActiveRecord::Migration
+  def change
+    add_column :variables, :labels, :text
+  end
+end